requirements.txt
Pillow>=10.0.0
watchdog==2.1.3
